Split into lib and app
